Put nav in its own component
